finishing manual mode
